small changes to the interface
